requirements.txt
Django==5.2.2
python-decouple==3.8
django-filter==25.1
djangorestframework==3.16.0
Markdown==3.8
djangorestframework-simplejwt[crypto]
django-redis==5.4.0
psycopg2==2.9.10
asgiref==3.8.1
cffi==1.17.1
cryptography==45.0.3
Django==5.2.2
django-cors-headers==4.7.0
django-filter==25.1
django-jazzmin==3.0.1
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
Markdown==3.8
packaging==25.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
redis==6.2.0
sqlparse==0.5.3
